+++ test/coverage/test_phase.py
@@ -1165,6 +1165,44 @@ def test_unspecified_http_methods(ctx):
     assert methods == {"HEAD", "PATCH", "TRACE", "DELETE", "OPTIONS", "PUT"}
 
 
+def test_no_missing_header_duplication(ctx):
+    schema = ctx.openapi.build_schema(
+        {
+            "/foo": {
+                "post": {
+                    "parameters": [
+                        {"name": "X-Key-1", "in": "header", "required": False, "schema": {"type": "string"}},
+                        {"name": "X-Key-2", "in": "header", "required": False, "schema": {"type": "string"}},
+                        {"name": "X-Key-3", "in": "header", "required": True, "schema": {"type": "string"}},
+                    ],
+                    "responses": {"default": {"description": "OK"}},
+                }
+            }
+        }
+    )
+    schema = schemathesis.from_dict(schema, validate_schema=True)
+
+    descriptions = []
+    operation = schema["/foo"]["post"]
+
+    def test(case):
+        if case.meta.phase != TestPhase.COVERAGE:
+            return
+        descriptions.append(case.meta.description)
+
+    test_func = create_test(
+        operation=operation,
+        test=test,
+        data_generation_methods=DataGenerationMethod.all(),
+        settings=settings(phases=[Phase.explicit]),
+    )
+
+    test_func()
+
+    assert "Missing required property: X-Key-3" not in descriptions
+    assert "Missing `X-Key-3` at header" in descriptions
+
+
 def assert_coverage(schema, methods, expected, path=None):
     schema = schemathesis.from_dict(schema, validate_schema=True)